cleanup of gene
